Remove PhysicalEmptyTree
A constant is enough.
