+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/DefaultFileSystemSnapshotter.java
@@ -27,7 +27,6 @@ import org.gradle.api.internal.changedetection.state.mirror.ImmutablePhysicalDir
 import org.gradle.api.internal.changedetection.state.mirror.MirrorUpdatingDirectoryWalker;
 import org.gradle.api.internal.changedetection.state.mirror.MutablePhysicalDirectorySnapshot;
 import org.gradle.api.internal.changedetection.state.mirror.MutablePhysicalSnapshot;
-import org.gradle.api.internal.changedetection.state.mirror.PhysicalEmptyTree;
 import org.gradle.api.internal.changedetection.state.mirror.PhysicalFileSnapshot;
 import org.gradle.api.internal.changedetection.state.mirror.PhysicalMissingSnapshot;
 import org.gradle.api.internal.changedetection.state.mirror.PhysicalSnapshot;
@@ -209,7 +208,7 @@ public class DefaultFileSystemSnapshotter implements FileSystemSnapshotter {
         if (rootSnapshot != null) {
             return rootSnapshot;
         }
-        return PhysicalEmptyTree.INSTANCE;
+        return PhysicalSnapshot.EMPTY;
     }
 
     private PhysicalSnapshot snapshotAndCache(DirectoryFileTree directoryTree) {
