+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/mirror/MirrorUpdatingDirectoryWalker.java
@@ -71,7 +71,7 @@ public class MirrorUpdatingDirectoryWalker {
     public PhysicalSnapshot walk(final PhysicalSnapshot fileSnapshot, @Nullable PatternSet patterns) {
         if (fileSnapshot.getType() == FileType.Missing) {
             // The root missing file should not be tracked for trees.
-            return PhysicalEmptyTree.INSTANCE;
+            return PhysicalSnapshot.EMPTY;
         }
         if (fileSnapshot.getType() == FileType.RegularFile) {
             return fileSnapshot;
